% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getUrl}
\alias{getUrl}
\title{Get the URL to which to connect to.}
\usage{
getUrl(host = NULL, port = 2908, useIP = TRUE, useFullName = FALSE)
}
\arguments{
\item{host}{the host of the URL - if \code{NULL} will be guessed.}

\item{port}{the port to connect to}

\item{useIP}{if \code{TRUE} will guess the IP instead of hostname. Connecting to IPs should be less hassle in most situations where
the device and the server are on the same subnet behind a NAT (e.g. at home) or behind a firewall (i.e. in a company).}

\item{useFullName}{get the hostname from System$getHostname.}
}
\value{
an URL string to connect to
}
\description{
Get the URL to which to connect to.
}
