% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotvr.R
\name{sendData}
\alias{sendData}
\title{Broadcast to a server via POST - hence this can run in another process or even on another host.}
\usage{
sendData(data, server = getOption("plotVR.broadcast.server",
  plotVR:::getUrl()), ...)
}
\arguments{
\item{data}{the JSON-model data to broadcast, from \link{plotVR}}

\item{server}{to which post to post the data, default also can be set
using \code{options(plotVR.broadcast.server="http://example.com:2908")}.}

\item{...}{passed to \code{httr::POST}.}
}
\value{
invisible the response of the server
}
\description{
Broadcast to a server via POST - hence this can run in another process or even on another host.
}
\examples{

\dontrun{
options(plotVF.default.broadcast=broadcastPost)
plotVR(iris)

# or directly:
plotVR(iris, broadcast=broadcastPost)
}
}
\seealso{
\code{\link{openController}}
}
