% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotvr.R
\name{startExternalServerProcess}
\alias{startExternalServerProcess}
\title{Start Server}
\usage{
startExternalServerProcess(...)
}
\arguments{
\item{...}{passed to \code{callr::r_bg}}
}
\value{
the process object (invisibly)
}
\description{
Start Server
}
